plugins/hybrid-kb-milvus-neo4j/agent-sdk/pyproject.toml
[project]
name = "knowledge-base-agent"
version = "1.0.0"
description = "Self-contained multi-tenant Knowledge Base Agent for Claude Agent SDK"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Arquitetos de Prompt", email = "gutomec@gmail.com"}
]
keywords = [
    "knowledge-base", "rag", "graphrag", "milvus", "neo4j",
    "claude", "agent-sdk", "multi-tenant", "embeddings"
]

dependencies = [
    "anthropic>=0.40.0",
    "pymilvus>=2.5.0",
    "neo4j>=5.0.0",
    "minio>=7.2.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "google-generativeai>=0.8.0",
    "cohere>=5.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
full = [
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
